+++ src/cmd/moby/gcp.go
